+++ corda-testacles-cordapp-example/src/test/kotlin/testacles/sample/test/mocknetwork/MockNetworkExtensionTest.kt
@@ -63,10 +63,12 @@ class MockNetworkExtensionTest {
                         names = CordaX500Names(listOf(
                                 CordaX500Name.parse("O=PartyA, L=Athens, C=GR"),
                                 CordaX500Name.parse("O=PartyB, L=Athens, C=GR"))),
-                        // Package names, one per cordapp to pickup
-                        cordappPackages = listOf<String>(
-                                SampleCordapp::class.java.`package`.name,
-                                PartitureFlow::class.java.`package`.name),
+                        // Optional, used *only* for the current
+                        // Gradle module, if a cordapp.
+                        cordappProjectPackage = SampleCordapp::class.java.`package`.name,
+                        // Optional; package names are used to pickup
+                        // cordapp or cordaCompile dependencies
+                        cordappPackages = listOf<String>(PartitureFlow::class.java.`package`.name),
                         // Optional, default
                         threadPerNode = true,
                         // Optional, default